Fix bug with checkout current step and login
When the step is the `login` step; Magento `checkout` instance is having `billing` as currentStep instead of `login`. Note that going the next step and going back to login actually fix it the current step is `login` as expected.

So we fix this bad behaviour by manually setting the correct current step if customer login is required.